+++ mugshot/src/main/java/uk/co/fractalmotion/mugshot/internal/differs/PixelPerfect.kt
@@ -39,30 +39,23 @@ internal object PixelPerfect : Differ {
         }
 
         if (expectedRgb == actualRgb) {
-          deltaImage.setRGB(expectedWidth + x, y, 0x00808080)
+          deltaImage.setRGB(expectedWidth + x, y, DeltaColors.UNCHANGED)
           continue
         }
 
         // If the pixels have no opacity, don't delta colors at all
         if (expectedRgb and -0x1000000 == 0 && actualRgb and -0x1000000 == 0) {
-          deltaImage.setRGB(expectedWidth + x, y, 0x00808080)
+          deltaImage.setRGB(expectedWidth + x, y, DeltaColors.UNCHANGED)
           continue
         }
 
         differentPixels++
 
         val deltaR = (actualRgb and 0xFF0000).ushr(16) - (expectedRgb and 0xFF0000).ushr(16)
-        val newR = 128 + deltaR and 0xFF
         val deltaG = (actualRgb and 0x00FF00).ushr(8) - (expectedRgb and 0x00FF00).ushr(8)
-        val newG = 128 + deltaG and 0xFF
         val deltaB = (actualRgb and 0x0000FF) - (expectedRgb and 0x0000FF)
-        val newB = 128 + deltaB and 0xFF
 
-        val avgAlpha =
-          ((expectedRgb and -0x1000000).ushr(24) + (actualRgb and -0x1000000).ushr(24)) / 2 shl 24
-
-        val newRGB = avgAlpha or (newR shl 16) or (newG shl 8) or newB
-        deltaImage.setRGB(expectedWidth + x, y, newRGB)
+        deltaImage.setRGB(expectedWidth + x, y, DeltaColors.DIFFERENT)
 
         delta += abs(deltaR).toLong()
         delta += abs(deltaG).toLong()